Reverse sendMessage handler and async method names
This is done to indicate that the handler only handles the request,
whereas the async method actually handles the processing.

This is necessary to hopefully prevent timeout issues for the related
endpoints since all I really want to do with this is process the data
coming in from Habitica. This service itself can then process the data
without slowing down Habitica which is probably just waiting for a
response.