No longer gzip content
This does not need to be done from this service for Discord to receive
the response. The issue on my end was that the error handler wasn't
being called, seemingly causing the connection to Discord to break and
a Caller Hung Up error to be logged. This was all fixed by binding the
proper `this` when using methods as handlers for promises.