Log data as JSON, ensure defaultMsg can be used
The first change is meant to make working with external logging easier.
The JSON will be more useful than when each part of the equivalent JS
object is logged on its own line. Papertrail in particular allows
expanding logged JSON in its new beta UI, and copying the JSON directly
will still be easier than copying and cleaning up multiple log lines.

The second change is meant to ensure that the default message is used
when `msg` is undefined. Before the function would throw an error and
halt because `.call` was being checked for on an undefined value. This
should be fixed now.