Reverse meaning of method name
Was returing the value as if the message type was found in the
blacklist, but was inverting the return value because I had named the
method `messageTypeAllowed`. The method is now called
`messageTypeDisallowed` since I'm checking a blacklist instead of a
whitelist, and the return value is no longer inverted. This is simpler
to reason about in this case anyway since we're dealing with a
blacklist.